Adjusting versions to get the site to build.
Getting eventmachine to work with ruby version 2 was a pain.
From what I could tell, eventmachine was referencing variables that were
not defined in the header files for newer versions of ruby.
Here are the changes that finally led to the successful creation of the
ruby environment with `bundle install` and then build of the website
with `bundle exec jekyll serve`:

-   Change the version of Ruby in the conda env to 2.4.1 (from 2.7.2).
-   Use bundler version 2.1.4 (installation command in README is updated
    to reflect this).
-   Roll back eventmachine version in Gemfile.lock from 1.2.7 to
    1.0.4. I didn't like having to edit the lock file for this all to
    work, but here we are.